% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGNoptim-package.R
\docType{package}
\name{RGNoptim-package}
\alias{RGNoptim}
\alias{RGNoptim-package}
\title{RGNoptim: R package: Robust Gauss-Newton Algorithm}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
