% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGNoptim.R
\name{RGNoptim2}
\alias{RGNoptim2}
\title{Robust Gauss-Newton Algorithm}
\usage{
RGNoptim2(x = c(0, 0), objective, residual, lower = NULL,
  upper = NULL, ...)
}
\description{
Robust Gauss-Newton Algorithm
}
\examples{
x = c(0, 0)
RGNoptim2(x, func_obj, func_res)
}
