% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGNoptim-package.R
\docType{package}
\name{RGNoptim-package}
\alias{RGNoptim}
\alias{RGNoptim-package}
\title{RGNoptim: R package: Robust Gauss-Newton Algorithm}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kongdd/RGNoptim}
  \item Report bugs at \url{https://github.com/kongdd/RGNoptim/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
